pass selectedSiteId as props to pricing meta hook
