+++ client/landing/stepper/declarative-flow/internals/steps-repository/pattern-assembler/screen-upsell.tsx
@@ -3,10 +3,12 @@ import { Button, Gridicon, LoadingPlaceholder } from '@automattic/components';
 import { formatCurrency } from '@automattic/format-currency';
 import { NavigatorHeader } from '@automattic/onboarding';
 import { useTranslate } from 'i18n-calypso';
+import { useSelector } from 'react-redux';
 import QueryPlans from 'calypso/components/data/query-plans';
 import useGlobalStylesUpgradeTranslations from 'calypso/components/premium-global-styles-upgrade-modal/use-global-styles-upgrade-translations';
 import PlanPrice from 'calypso/my-sites/plan-price';
 import usePricingMetaForGridPlans from 'calypso/my-sites/plans-features-main/hooks/data-store/use-pricing-meta-for-grid-plans';
+import { getSelectedSiteId } from 'calypso/state/ui/selectors';
 import { useScreen } from './hooks';
 import NavigatorTitle from './navigator-title';
 import './screen-upsell.scss';
@@ -21,9 +23,13 @@ const ScreenUpsell = ( { numOfSelectedGlobalStyles = 1, onCheckout, onTryStyle }
 	const translate = useTranslate();
 	const { title, description } = useScreen( 'upsell' );
 	const translations = useGlobalStylesUpgradeTranslations( { numOfSelectedGlobalStyles } );
+	const selectedSiteId = useSelector( getSelectedSiteId ) ?? undefined;
+
+	// TODO clk pricing
 	const pricingMeta = usePricingMetaForGridPlans( {
 		planSlugs: [ PLAN_PREMIUM ],
 		storageAddOns: null,
+		selectedSiteId,
 	} );
 
 	const pricing = pricingMeta?.[ PLAN_PREMIUM ];
