+++ client/my-sites/plans-features-main/hooks/data-store/use-pricing-meta-for-grid-plans.ts
@@ -4,45 +4,13 @@ import {
 	getTermFromDuration,
 	calculateMonthlyPrice,
 } from '@automattic/calypso-products';
-import { Plans, WpcomPlansUI, Purchases, type AddOnMeta } from '@automattic/data-stores';
+import { Plans, WpcomPlansUI, Purchases } from '@automattic/data-stores';
 import { useSelect } from '@wordpress/data';
-import { useSelector } from 'react-redux';
-import getSelectedSiteId from 'calypso/state/ui/selectors/get-selected-site-id';
 import useCheckPlanAvailabilityForPurchase from '../use-check-plan-availability-for-purchase';
-
-export interface PricingMetaForGridPlan {
-	billingPeriod?: Plans.PlanPricing[ 'billPeriod' ];
-	currencyCode?: Plans.PlanPricing[ 'currencyCode' ];
-	originalPrice: Plans.PlanPricing[ 'originalPrice' ];
-	/**
-	 * If discounted prices are provided (not null), they will take precedence over originalPrice.
-	 * UI will show original with a strikethrough or grayed out
-	 */
-	discountedPrice: Plans.PlanPricing[ 'discountedPrice' ];
-	/**
-	 * Intro offers override billing and pricing shown in the UI
-	 * they are currently defined off the site plans (so not defined when siteId is not available)
-	 */
-	introOffer?: Plans.PlanPricing[ 'introOffer' ];
-	expiry?: Plans.SitePlan[ 'expiry' ];
-}
-
-export type UsePricingMetaForGridPlans = ( {
-	planSlugs,
-	withoutProRatedCredits,
-	storageAddOns,
-}: {
-	planSlugs: PlanSlug[];
-	withoutProRatedCredits?: boolean;
-	storageAddOns: ( AddOnMeta | null )[] | null;
-} ) => { [ planSlug: string ]: PricingMetaForGridPlan } | null;
-
-interface Props {
-	planSlugs: PlanSlug[];
-	withoutProRatedCredits?: boolean;
-	storageAddOns?: ( AddOnMeta | null )[] | null;
-	coupon?: string;
-}
+import type {
+	UsePricingMetaForGridPlans,
+	PricingMetaForGridPlan,
+} from 'calypso/my-sites/plans-grid/hooks/npm-ready/data-store/use-grid-plans';
 
 function getTotalPrice( planPrice: number | null | undefined, addOnPrice = 0 ): number | null {
 	return null !== planPrice && undefined !== planPrice ? planPrice + addOnPrice : null;
@@ -56,10 +24,9 @@ const usePricingMetaForGridPlans: UsePricingMetaForGridPlans = ( {
 	planSlugs,
 	withoutProRatedCredits = false,
 	storageAddOns,
+	selectedSiteId,
 	coupon,
-}: Props ) => {
-	// TODO: pass this in as a prop to uncouple the dependency
-	const selectedSiteId = useSelector( getSelectedSiteId ) ?? undefined;
+} ) => {
 	// TODO: pass this in as a prop to uncouple the dependency
 	const planAvailabilityForPurchase = useCheckPlanAvailabilityForPurchase( { planSlugs } );
