+++ client/my-sites/plans-features-main/hooks/test/use-pricing-meta-for-grid-plans.ts
@@ -26,14 +26,12 @@ jest.mock( '../use-check-plan-availability-for-purchase', () => jest.fn() );
 
 import { PLAN_PERSONAL, PLAN_PREMIUM } from '@automattic/calypso-products';
 import { Plans, Purchases } from '@automattic/data-stores';
-import getSelectedSiteId from 'calypso/state/ui/selectors/get-selected-site-id';
 import usePricingMetaForGridPlans from '../data-store/use-pricing-meta-for-grid-plans';
 import useCheckPlanAvailabilityForPurchase from '../use-check-plan-availability-for-purchase';
 
 describe( 'usePricingMetaForGridPlans', () => {
 	beforeEach( () => {
 		jest.clearAllMocks();
-		getSelectedSiteId.mockImplementation( () => 100 );
 		Purchases.useSitePurchaseById.mockImplementation( () => undefined );
 		Plans.useIntroOffers.mockImplementation( () => ( {
 			[ PLAN_PREMIUM ]: null,
@@ -95,6 +93,7 @@ describe( 'usePricingMetaForGridPlans', () => {
 			planSlugs: [ PLAN_PREMIUM ],
 			withoutProRatedCredits: false,
 			storageAddOns: null,
+			selectedSiteId: 100,
 		} );
 
 		const expectedPricingMeta = {
@@ -132,6 +131,7 @@ describe( 'usePricingMetaForGridPlans', () => {
 			planSlugs: [ PLAN_PREMIUM ],
 			withoutProRatedCredits: false,
 			storageAddOns: null,
+			selectedSiteId: 100,
 		} );
 
 		const expectedPricingMeta = {
@@ -170,6 +170,7 @@ describe( 'usePricingMetaForGridPlans', () => {
 			planSlugs: [ PLAN_PREMIUM ],
 			withoutProRatedCredits: false,
 			storageAddOns: null,
+			selectedSiteId: 100,
 		} );
 
 		const expectedPricingMeta = {
@@ -208,6 +209,7 @@ describe( 'usePricingMetaForGridPlans', () => {
 			planSlugs: [ PLAN_PREMIUM ],
 			withoutProRatedCredits: true,
 			storageAddOns: null,
+			selectedSiteId: 100,
 		} );
 
 		const expectedPricingMeta = {
