+++ client/my-sites/plans-grid/components/plan-type-selector/components/interval-type-toggle.tsx
@@ -21,18 +21,21 @@ export const IntervalTypeToggle: React.FunctionComponent< IntervalTypeProps > = 
 		usePricingMetaForGridPlans,
 		title,
 		coupon,
+		selectedSiteId,
 	} = props;
 	const [ spanRef, setSpanRef ] = useState< HTMLSpanElement >();
 	const segmentClasses = classNames( 'price-toggle', {
 		'is-signup': isInSignup,
 	} );
 	const popupIsVisible = Boolean( intervalType === 'monthly' && isInSignup && props.plans.length );
-	const maxDiscount = useMaxDiscount( props.plans, usePricingMetaForGridPlans );
+	const maxDiscount = useMaxDiscount( props.plans, usePricingMetaForGridPlans, selectedSiteId );
+	// TODO clk pricing
 	const pricingMeta = usePricingMetaForGridPlans( {
 		planSlugs: currentSitePlanSlug ? [ currentSitePlanSlug ] : [],
 		withoutProRatedCredits: true,
 		storageAddOns: null,
 		coupon,
+		selectedSiteId,
 	} );
 	const currentPlanBillingPeriod = currentSitePlanSlug
 		? pricingMeta?.[ currentSitePlanSlug ]?.billingPeriod
