+++ client/my-sites/plans-grid/index.tsx
@@ -11,7 +11,7 @@ import './style.scss';
 import type { ComparisonGridExternalProps, FeaturesGridExternalProps } from './types';
 
 const WrappedComparisonGrid = ( {
-	siteId,
+	selectedSiteId,
 	intent,
 	gridPlans,
 	usePricingMetaForGridPlans,
@@ -37,6 +37,7 @@ const WrappedComparisonGrid = ( {
 	return (
 		<PlansGridContextProvider
 			intent={ intent }
+			selectedSiteId={ selectedSiteId }
 			gridPlans={ gridPlans }
 			usePricingMetaForGridPlans={ usePricingMetaForGridPlans }
 			allFeaturesList={ allFeaturesList }
@@ -48,7 +49,7 @@ const WrappedComparisonGrid = ( {
 				isLaunchPage={ isLaunchPage }
 				currentSitePlanSlug={ currentSitePlanSlug }
 				onUpgradeClick={ handleUpgradeClick }
-				siteId={ siteId }
+				selectedSiteId={ selectedSiteId }
 				selectedPlan={ selectedPlan }
 				selectedFeature={ selectedFeature }
 				showUpgradeableStorage={ showUpgradeableStorage }
@@ -62,7 +63,7 @@ const WrappedComparisonGrid = ( {
 
 const WrappedFeaturesGrid = ( props: FeaturesGridExternalProps ) => {
 	const {
-		siteId,
+		selectedSiteId,
 		intent,
 		gridPlans,
 		usePricingMetaForGridPlans,
@@ -72,7 +73,7 @@ const WrappedFeaturesGrid = ( props: FeaturesGridExternalProps ) => {
 	} = props;
 	const translate = useTranslate();
 	const isPlanUpgradeCreditEligible = useIsPlanUpgradeCreditVisible(
-		siteId,
+		selectedSiteId,
 		gridPlans.map( ( gridPlan ) => gridPlan.planSlug )
 	);
 	const { prices, currencyCode } = usePlanPricingInfoFromGridPlans( {
@@ -91,6 +92,7 @@ const WrappedFeaturesGrid = ( props: FeaturesGridExternalProps ) => {
 	return (
 		<PlansGridContextProvider
 			intent={ intent }
+			selectedSiteId={ selectedSiteId }
 			gridPlans={ gridPlans }
 			usePricingMetaForGridPlans={ usePricingMetaForGridPlans }
 			coupon={ coupon }
